# docs/05-api/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory API
  version: 1.0.0
  description: API interna y pública para gestión de inventario StockFlow.
  contact:
    name: StockFlow Support
    email: soporte@tuempresa.com
servers:
  - url: https://inventario.tuempresa.com
    description: Production
  - url: http://localhost:8080
    description: Local backend
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: Listar todos los usuarios
      description: Obtiene una lista paginada de todos los usuarios registrados en el sistema. (Solo administradores).
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
      responses:
        '200':
          description: Lista de usuarios recuperada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUserResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Permisos insuficientes (Requiere rol ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags: [Users]
      summary: Registrar nuevo usuario
      description: Registra un nuevo usuario en el sistema. (Solo administradores).
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos de entrada no válidos o conflicto de datos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Email ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/offices:
    get:
      tags: [Offices]
      summary: Listar todas las oficinas
      description: Recupera una lista paginada de las sedes de la empresa.
      operationId: listOffices
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
      responses:
        '200':
          description: Lista de oficinas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfficeResponse'
    post:
      tags: [Offices]
      summary: Crear oficina
      description: Registra una nueva sede en el sistema.
      operationId: createOffice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeCreateRequest'
      responses:
        '201':
          description: Sede registrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/offices/{publicId}:
    put:
      tags: [Offices]
      summary: Actualizar nombre de oficina
      description: Modifica los parámetros descriptivos de una oficina existente.
      operationId: updateOffice
      security:
        - bearerAuth: []
      parameters:
        - name: publicId
          in: path
          required: true
          description: Identificador público de la sede.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfficeCreateRequest'
      responses:
        '200':
          description: Oficina modificada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfficeResponse'
        '404':
          description: Sede no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/materials:
    get:
      tags: [Materials]
      summary: Listar e investigar materiales de inventario
      description: Devuelve una lista de materiales según los filtros opcionales aplicados.
      operationId: listMaterials
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por estado físico.
          required: false
          schema:
            type: string
            enum: [OPERATIVO, ROTO, EN_REPARACION, BAJA]
        - name: officePublicId
          in: query
          description: Filtrar por identificador de oficina asociada.
          required: false
          schema:
            type: string
        - name: materialType
          in: query
          description: Filtrar por tipo o categoría de equipo.
          required: false
          schema:
            type: string
        - name: serialNumber
          in: query
          description: Filtrar por coincidencia parcial de número de serie.
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
      responses:
        '200':
          description: Materiales encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageMaterialResponse'
    post:
      tags: [Materials]
      summary: Registrar nuevo material
      description: Da de alta un nuevo artículo, le asocia una oficina y auto-genera su código público y código QR único de inventario.
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '201':
          description: Material creado y registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/materials/{publicCode}/qr:
    get:
      tags: [Inventory]
      summary: Recuperar imagen del código QR
      description: Devuelve la imagen en formato PNG que contiene el código QR de trazabilidad rápida del material.
      operationId: getMaterialQr
      security:
        - bearerAuth: []
      parameters:
        - name: publicCode
          in: path
          required: true
          description: Código único seguro de inventario (mat_XXXXXXXX).
          schema:
            type: string
      responses:
        '200':
          description: Imagen del QR binaria en PNG
          content:
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Material o código QR no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/materials/{publicCode}/qr/regenerate:
    post:
      tags: [Inventory]
      summary: Regenerar código QR del material
      description: Fuerza la recreación del código QR del artículo. Útil si se corrompió el archivo de almacenamiento.
      operationId: regenerateMaterialQr
      security:
        - bearerAuth: []
      parameters:
        - name: publicCode
          in: path
          required: true
          description: Código único seguro de inventario (mat_XXXXXXXX).
          schema:
            type: string
      responses:
        '204':
          description: Código QR regenerado exitosamente
        '404':
          description: Material no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Introducir el Access Token JWT precedido del prefijo 'Bearer '.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de acceso estática para integradores externos autorizados.
  parameters:
    pageParam:
      name: page
      in: query
      description: Número de página a consultar (base 0).
      required: false
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      in: query
      description: Cantidad máxima de registros por página.
      required: false
      schema:
        type: integer
        default: 10
  schemas:
...
    UserResponse:
      type: object
      properties:
        publicId:
          type: string
          format: uuid
          example: e2c92a95-5dbd-4a18-912c-354f9a0a03bb
        email:
          type: string
          format: email
          example: tecnico@tuempresa.com
        role:
          type: string
          enum: [ADMIN, TECNICO]
          example: TECNICO
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-22T10:30:00Z"
    UserCreateRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
          example: nuevo.tecnico@tuempresa.com
        password:
          type: string
          minLength: 8
          example: tEcNiCo2026!
        role:
          type: string
          enum: [ADMIN, TECNICO]
          example: TECNICO
...
    OfficeResponse:
      type: object
      properties:
        publicId:
          type: string
          example: off-c8d7e6f5
        name:
          type: string
          example: Sede Central Madrid
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
    OfficeCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Oficina de Getafe
    MaterialResponse:
      type: object
      properties:
        publicCode:
          type: string
          example: mat_a1b2c3d4
        materialType:
          type: string
          example: Router
        brand:
          type: string
          example: MikroTik
        model:
          type: string
          example: RB4011
        serialNumber:
          type: string
          example: SN_MIKRO_98765
        description:
          type: string
          example: Router troncal de comunicación principal
        status:
          type: string
          enum: [OPERATIVO, ROTO, EN_REPARACION, BAJA]
          example: OPERATIVO
        officePublicId:
          type: string
          example: off-c8d7e6f5
        officeName:
          type: string
          example: Sede Central Madrid
        createdByName:
          type: string
          example: Juan Pérez
        updatedByName:
          type: string
          example: Juan Pérez
        createdAt:
          type: string
          format: date-time
          example: "2026-05-22T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-22T08:30:00Z"
    MaterialCreateRequest:
      type: object
      required: [materialType, brand, model, serialNumber, officePublicId]
      properties:
        materialType:
          type: string
          example: Switch
        brand:
          type: string
          example: Cisco
        model:
          type: string
          example: Catalyst 2960L
        serialNumber:
          type: string
          example: SN_CISCO_445566
        description:
          type: string
          example: Switch PoE de 24 puertos
        officePublicId:
          type: string
          example: off-c8d7e6f5
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: La justificación es obligatoria para dar de baja un equipo.
        timestamp:
          type: string
          format: date-time
          example: "2026-05-22T08:30:01Z"
    PageUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 5
        totalPages:
          type: integer
          example: 1
        last:
          type: boolean
          example: true
    PageOfficeResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OfficeResponse'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 3
        totalPages:
          type: integer
          example: 1
        last:
          type: boolean
          example: true
    PageMaterialResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/MaterialResponse'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        totalElements:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        last:
          type: boolean
          example: false
...